Rename quantize_wrapper to fully_quantized_quantize_wrapper
